requirements.txt
googletrans==4.0.0-rc1
gtts==2.3.0
google-generativeai
streamlit
python-dotenv
Pillow
python-docx
PyMuPDF
pydrive2
pypdf
langchain
torch
accelerate
bitsandbytes
ctransformers
sentence_transformers
faiss_cpu
chainlit
huggingface_hub
langchain_community